Add LockfileV2 and Resolvers modules
